+++ re-common/src/main/java/cn/lingjiatong/re/common/constant/UserConstant.java
@@ -11,7 +11,7 @@ public interface UserConstant {
     // 超级管理员用户名
     String SUPER_ADMIN_USER = "lingjiatong";
     // 超级管理员ID
-    Integer SUPER_ADMIN_USER_ID = 1;
+    Long SUPER_ADMIN_USER_ID = 1L;
     // 系统操作用户名
     String SYSTEM_USER = "system";
 }
